src/components/AppBar/hooks/useAppBar.ts
import { MouseEvent, useCallback, useState } from "react";

const useAppBar = () => {
  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
  const isMenuOpen = Boolean(anchorEl);

  const handleOpenMenu = useCallback(
    (event: MouseEvent<HTMLButtonElement>) => {
      setAnchorEl(event.currentTarget);
    },
    [setAnchorEl],
  );

  const handleCloseMenu = useCallback(() => {
    setAnchorEl(null);
  }, [setAnchorEl]);

  return { anchorEl, isMenuOpen, handleOpenMenu, handleCloseMenu };
};

export default useAppBar;
